Fix vkWholeSize

+++ vulkan.nim
@@ -405,7 +405,7 @@ const
   vkLodClampNone* = 1000.0
   vkRemainingMipLevels* = (not 0'u32)
   vkRemainingArrayLayers* = (not 0'u32)
-  vkWholeSize* = (not 0'u32)
+  vkWholeSize* = (not 0'u64)
   vkAttachmentUnused* = (not 0'u32)
   vkTrue* = 1
   vkFalse* = 0
